:sparkles: Use esbuild directly for produce final build
Replace the vite abstraction around esbuild because
it does not allows properly minify the file and remove all
docstring comments from it.

For some reasons, vite decide that is very nice to disable
the ability to strip comments and minify when the target is
a library. Vite internally uses esbuild, so we just skiped
the vite abstraction and use esbuild directly for produce
the final bundle.